Refactor metrics tests structure
