+++ pycktool/tests/metrics/base_metrics/test_logical_lines_of_code.py
@@ -1,0 +1,3 @@
+
+class TestLogicalLinesOfCode:
+    pass
