+++ pycktool/tests/metrics/class_metrics/test_lack_of_cohesion.py
@@ -1,0 +1,4 @@
+import pytest
+
+class TestLackOfCohesion:
+    pass
