+++ pycktool/tests/metrics/class_metrics/test_number_of_attributes.py
@@ -1,0 +1,4 @@
+import pytest
+
+class TestNumberOfAttributes:
+    pass
