+++ pycktool/tests/metrics/class_metrics/test_number_of_methods.py
@@ -1,0 +1,4 @@
+import pytest
+
+class TestNumberOfMethods:
+    pass
