+++ pycktool/tests/metrics/class_metrics/test_response_for_class.py
@@ -1,0 +1,4 @@
+import pytest
+
+class TestResponseForClass:
+    pass
