+++ pycktool/tests/metrics/method_metrics/test_number_of_parameters.py
@@ -1,0 +1,4 @@
+import pytest
+
+class TestNumberOfParameters:
+    pass
